[Multimedia] Moved enums of MediaTool into separate files.
